Revert "seperate main binary into executable and library (#846)"
This reverts commit 3d10d4a8722b6ad350ecc0f57580cc445cd9965b.
