+++ src/app/medInria/CMakeLists.txt
@@ -103,19 +103,10 @@ if (WIN32)
 endif()
 
 ## #############################################################################
-## Add targets
+## Add Exe
 ## #############################################################################
 
 add_executable(${TARGET_NAME} ${DEPLOYMENT_SYSTEM} # Empty for Linux
-    main.cpp
-  )
-
-if (APPLE)
-  set_target_properties(${TARGET_NAME} PROPERTIES MACOSX_BUNDLE_INFO_PLIST "${CMAKE_SOURCE_DIR}/cmake/MedInriaOSXBundleInfo.plist.in")
-  import_external_resources(${TARGET_NAME} FILES "${CMAKE_CURRENT_SOURCE_DIR}/resources/${TARGET_NAME}.icns")
-endif()
-
-add_library(${TARGET_NAME}_LIBRARY
   ${${TARGET_NAME}_CFILES}
   ${${TARGET_NAME}_QRC}
   ${${TARGET_NAME}_QSS} # Just to have it in EDI
@@ -123,13 +114,16 @@ add_library(${TARGET_NAME}_LIBRARY
   ${${TARGET_NAME}_HTML}
   )
 
-set_target_properties(${TARGET_NAME}_LIBRARY PROPERTIES OUTPUT_NAME "${PROJECT_NAME}")
+if (APPLE)
+  set_target_properties(${TARGET_NAME} PROPERTIES MACOSX_BUNDLE_INFO_PLIST "${CMAKE_SOURCE_DIR}/cmake/MedInriaOSXBundleInfo.plist.in")
+  import_external_resources(${TARGET_NAME} FILES "${CMAKE_CURRENT_SOURCE_DIR}/resources/${TARGET_NAME}.icns")
+endif()
 
 ## #############################################################################
 ## include directorie.
 ## #############################################################################
 
-target_include_directories(${TARGET_NAME}_LIBRARY
+target_include_directories(${TARGET_NAME}
   PRIVATE ${${TARGET_NAME}_INCLUDE_DIRS}
   ${Boost_INCLUDE_DIR}
   )
@@ -140,10 +134,6 @@ target_include_directories(${TARGET_NAME}_LIBRARY
 ## #############################################################################
 
 target_link_libraries(${TARGET_NAME}
-    ${TARGET_NAME}_LIBRARY
-    )
-
-target_link_libraries(${TARGET_NAME}_LIBRARY
   Qt5::Core
   Qt5::Widgets
   Qt5::Gui
@@ -164,7 +154,7 @@ target_link_libraries(${TARGET_NAME}_LIBRARY
 ## #############################################################################
 
 if(USE_Python)
-    target_link_libraries(${TARGET_NAME}_LIBRARY ${pyncpp_CPP_API_LIBRARY})
+    target_link_libraries(${TARGET_NAME} ${pyncpp_CPP_API_LIBRARY})
 
     if(APPLE)
         set(python_home "../Resources")
@@ -180,7 +170,7 @@ if(USE_Python)
         set(python_plugin_path "python_plugins")
     endif()
 
-    target_compile_definitions(${TARGET_NAME}_LIBRARY PUBLIC
+    target_compile_definitions(${TARGET_NAME} PUBLIC
         USE_PYTHON
         PYTHON_HOME="${python_home}"
         PYTHON_PLUGIN_PATH="${python_plugin_path}"
@@ -189,7 +179,7 @@ if(USE_Python)
         )
 
     if(WIN32)
-        target_compile_definitions(${TARGET_NAME}_LIBRARY PUBLIC
+        target_compile_definitions(${TARGET_NAME} PUBLIC
             PLUGINS_LEGACY_PATH="plugins_legacy"
             )
     endif()
@@ -201,4 +191,3 @@ endif()
 
 ## see cmake/module/set_lib_install_rules.Cmake
 set_exe_install_rules(${TARGET_NAME})
-set_lib_install_rules(${TARGET_NAME}_LIBRARY)
