--- src/app/medInria/medInria.h
@@ -1,17 +1,0 @@
-#pragma once
-/*=========================================================================
-
- medInria
-
- Copyright (c) INRIA 2024. All rights reserved.
- See LICENSE.txt for details.
-
-  This software is distributed WITHOUT ANY WARRANTY; without even
-  the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
-  PURPOSE.
-
-=========================================================================*/
-
-#include "medInriaExport.h"
-
-MEDINRIA_EXPORT int run(int argc,char* argv[]);
